Moving components on the folders
